+++ lib/widgets/google_map_places_auto_complete.dart
@@ -13,9 +13,15 @@ class GoogleMapPlacesAutoComplete{
 
 
 
-  static Future<List<Prediction>> getPlaces(String placeName, String mapsKey) async {
+  static Future<List<Prediction>> getPlaces(String placeName, String mapsKey,{String? otherOptions}) async {
     /// referenceUrl = https://developers.google.com/maps/documentation/places/web-service/autocomplete#maps_http_places_autocomplete_amoeba-txt;
-    http.Response response  = await http.get(Uri.parse("https://maps.googleapis.com/maps/api/place/autocomplete/json?input=$placeName&key=$mapsKey"));
+    String mapOptions = [
+      'input=$placeName',
+      'key=$mapsKey',
+      otherOptions
+    ].join('&');
+
+    http.Response response  = await http.get(Uri.parse("https://maps.googleapis.com/maps/api/place/autocomplete/json?$mapOptions"));
     debugPrint(response.body.toString());
     if(response.statusCode == 200){
       Predictions predictions = predictionsFromJson(response.body);
@@ -27,11 +33,16 @@ class GoogleMapPlacesAutoComplete{
     }
     return [];
   }
-  static Future<Place?> getPlaceDetail(String placeId, String mapsKey) async {
+  static Future<Place?> getPlaceDetail(String placeId, String mapsKey, {String? otherOptions}) async {
     /// referenceUrl = https://developers.google.com/maps/documentation/places/web-service/place-id
     // https://maps.googleapis.com/maps/api/place/details/json?place_id=ChIJrTLr-GyuEmsRBfy61i59si0&key=YOUR_API_KEY
+    String mapOptions = [
+      'place_id=$placeId',
+      'key=$mapsKey',
+      otherOptions
+    ].join('&');
 
-    http.Response response  = await http.get(Uri.parse("https://maps.googleapis.com/maps/api/place/details/json?place_id=$placeId&key=$mapsKey"));
+    http.Response response  = await http.get(Uri.parse("https://maps.googleapis.com/maps/api/place/details/json?$mapOptions"));
     debugPrint(response.body.toString());
     if(response.statusCode == 200){
       return placeFromJson(response.body);